Update crates
- Refactor templates
- Add errors
